Demonstrate nework works
Despite my comment in the last commit, the network code was actually
functioning as expected. My error was in Network.classify, where I was
searching for the index of the maximum value in the array of output neurons, as
opposed to the array of neuron values.

I also properly parameterized the learning rate variables in this commit